fixing formatting of number groupings not working correctly
